# store-monitor/takealot-marketplace-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Takealot Marketplace API（卖家市场 API）",
    "version": "1.0.0",
    "description": "Takealot Marketplace API v1.0.0 — 卖家市场 REST 接口。\n\n## 鉴权\n- **大部分接口 必填**：Header `X-API-Key: <密钥>`（卖家后台 API Access 生成）\n- **无需鉴权**：GET /status\n- 若 403 且返回 HTML 页面，请添加浏览器 Header（User-Agent 等），见公共 Header\n\n## 参数标记说明\n- **[必填]**：必须提供，否则请求失败或无法定位资源\n- **[可选]**：可不传；不传时使用平台默认值或返回全部可匹配数据\n- **[建议]**：非强制，但强烈建议填写（如防 Cloudflare 拦截的 Header）\n\n## 重要限制\n- Shipments（发货）模块**只读**，无 POST 创建发货单接口\n- 带日期的列表接口单次查询跨度最多 **180 天**\n- 分页：使用响应中的 continuation_token；携带 token 时其它参数会被忽略\n- 创建报价 POST /offers 要求商品**已存在于 Takealot 目录**"
  },
  "servers": [
    {
      "url": "https://marketplace-api.takealot.com/v1",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/returns": {
      "get": {
        "tags": [
          "Returns"
        ],
        "summary": "获取退货列表",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserAgent"
          },
          {
            "$ref": "#/components/parameters/AcceptJson"
          },
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false,
            "description": "[可选] 仅返回指定字段，减小响应体积。Offers 可用：offer_id, tsin_id, sku, barcode, product_label, selling_price, rrp, minimum_leadtime_days, status, title, discount_percentage, storage_fee_eligible, created_at, updated_at, affected_by_vacation, disabled_by_seller, disabled_by_takealot, conversion_percentage_30_days, conversion_percentage_previous_30_days, page_views_30_days, quantity_returned_30_days, condition, image_url, productline_id, width_cm, length_cm, height_cm, weight_grams, is_conveyable, benchmark_price, total_wishlist, wishlist_30_days, listing_quality"
          },
          {
            "name": "expands",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "outcomes",
                  "transactions"
                ]
              }
            },
            "required": false,
            "description": "[可选] 展开关联数据。Offers 可用：seller_warehouse_stock, offer_charges, takealot_warehouse_stock, replenishment_blocks。replenishment_blocks 仅在报价被完全补货封锁时才有数据。"
          },
          {
            "name": "return_reason__in",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "changed_my_mind",
                  "customer_cancellation",
                  "defective_or_damaged",
                  "exception",
                  "exchange_for_a_different_size_colour",
                  "failed_delivery",
                  "not_what_i_ordered"
                ]
              }
            },
            "required": false,
            "description": "[可选] 退货原因。changed_my_mind | customer_cancellation | defective_or_damaged | exception | exchange_for_a_different_size_colour | failed_delivery | not_what_i_ordered"
          },
          {
            "name": "outcome__in",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "pending_removal_order",
                  "removal_order",
                  "pending_sellable_stock",
                  "sellable_stock"
                ]
              }
            },
            "required": false,
            "description": "[可选] 退货处理结果。pending_removal_order | removal_order | pending_sellable_stock | sellable_stock"
          },
          {
            "name": "return_region__in",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "CPT",
                  "JHB",
                  "DBN",
                  "cpt",
                  "jhb",
                  "dbn"
                ]
              }
            },
            "required": false,
            "description": "[可选] 退货处理区域。CPT/JHB/DBN（大小写均可）。"
          },
          {
            "name": "return_date__gte",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2025-01-01",
            "required": false,
            "description": "[可选] 退货日期起始（含）。"
          },
          {
            "name": "return_date__lte",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2025-12-31",
            "required": false,
            "description": "[可选] 退货日期截止（含）。"
          },
          {
            "name": "order_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": 9876543,
            "required": false,
            "description": "[可选] 按 Takealot 客户订单 ID 过滤。integer ≥1。"
          },
          {
            "name": "return_reference_number",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "RRN-QBYQ7-E67K",
            "required": false,
            "description": "[可选] 客户退货参考号 RRN，如 RRN-QBYQ7-E67K。"
          },
          {
            "name": "sku",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "SKU-ABC123",
            "required": false,
            "description": "[可选] 按商家 SKU 过滤。"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "required": false,
            "description": "[可选] 每页条数。Offers/Shipments 默认 100、最大 1000；Sales/Transactions 默认 20、最大 100；Returns 默认 10、最大 100。"
          },
          {
            "name": "continuation_token",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "[可选] 翻页令牌。从上一页响应获取；携带时其它查询参数会被忽略，且不返回 count。"
          },
          {
            "name": "order_by",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "return_date",
                "-return_date"
              ],
              "default": "-return_date"
            },
            "required": false,
            "description": "[可选] 排序字段。前缀 - 表示降序。Offers 默认 offer_id；Shipments 默认 -shipment_id；Returns 默认 -return_date。"
          },
          {
            "name": "include_count",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "required": false,
            "description": "[可选] 是否返回符合条件的总条数 count。默认 false；设为 true 可能影响性能。"
          }
        ],
        "responses": {
          "200": {
            "description": "退货列表",
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "seller_return_id": 12345,
                      "order_id": 9876543,
                      "offer_id": 123456,
                      "return_reference_number": "RRN-QBYQ7-E67K",
                      "sku": "SKU-ABC123",
                      "return_region": "JHB",
                      "quantity": 1,
                      "return_date": "2024-05-15",
                      "return_reason": "defective_or_damaged"
                    }
                  ],
                  "count": 150,
                  "limit": 100,
                  "continuation_token": "eyJvZmZzZXQiOiAxMDB9"
                }
              }
            }
          },
          "403": {
            "description": "API key 缺失或无效"
          }
        },
        "description": "分页查询卖家退货记录。**鉴权必填**。所有 Query 均为可选；expands=outcomes,transactions 可展开处理结果与关联交易。"
      }
    },
    "/returns/{return_id}": {
      "get": {
        "tags": [
          "Returns"
        ],
        "summary": "按 ID 获取退货",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserAgent"
          },
          {
            "$ref": "#/components/parameters/AcceptJson"
          },
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "name": "return_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 9876543,
            "description": "[必填] 卖家退货 ID（seller_return_id）。"
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false,
            "description": "[可选] 仅返回指定字段，减小响应体积。Offers 可用：offer_id, tsin_id, sku, barcode, product_label, selling_price, rrp, minimum_leadtime_days, status, title, discount_percentage, storage_fee_eligible, created_at, updated_at, affected_by_vacation, disabled_by_seller, disabled_by_takealot, conversion_percentage_30_days, conversion_percentage_previous_30_days, page_views_30_days, quantity_returned_30_days, condition, image_url, productline_id, width_cm, length_cm, height_cm, weight_grams, is_conveyable, benchmark_price, total_wishlist, wishlist_30_days, listing_quality"
          },
          {
            "name": "expands",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "outcomes",
                  "transactions"
                ]
              }
            },
            "required": false,
            "description": "[可选] 展开关联数据。Offers 可用：seller_warehouse_stock, offer_charges, takealot_warehouse_stock, replenishment_blocks。replenishment_blocks 仅在报价被完全补货封锁时才有数据。"
          }
        ],
        "responses": {
          "200": {
            "description": "单个退货"
          },
          "403": {
            "description": "API key 缺失或无效"
          },
          "404": {
            "description": "未找到"
          }
        },
        "description": "按 ID 获取单条退货详情。**鉴权必填**。路径 return_id 必填。"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "[必填] 卖家 API 密钥。在 Seller Portal → API Integrations → Authentication 生成。Header 名：X-API-Key。每个卖家账户同时仅 1 个有效 Key。若无效返回 403 JSON（非 HTML）。"
      }
    },
    "parameters": {
      "UserAgent": {
        "name": "User-Agent",
        "in": "header",
        "required": false,
        "description": "[建议] 模拟浏览器 User-Agent，避免 Cloudflare 返回 403 HTML 拦截页。Apifox/脚本调用时建议始终带上。",
        "schema": {
          "type": "string",
          "default": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
        },
        "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
      },
      "AcceptJson": {
        "name": "Accept",
        "in": "header",
        "required": false,
        "description": "[建议] 声明期望 JSON 响应，建议值为 application/json。",
        "schema": {
          "type": "string",
          "default": "application/json"
        },
        "example": "application/json"
      },
      "AcceptLanguage": {
        "name": "Accept-Language",
        "in": "header",
        "required": false,
        "description": "[可选] 语言偏好，不影响 API 业务逻辑。",
        "schema": {
          "type": "string",
          "default": "zh-CN,zh;q=0.9,en;q=0.8"
        },
        "example": "zh-CN,zh;q=0.9,en;q=0.8"
      }
    },
...
  },
...
}